+++ cpp/src/strings/sorting/sorting.cu
@@ -21,6 +21,8 @@
 #include <cudf/strings/strings_column_view.hpp>
 
 #include <rmm/thrust_rmm_allocator.h>
+#include <rmm/cuda_stream_view.hpp>
+
 #include <thrust/sequence.h>
 #include <thrust/sort.h>
 
@@ -32,7 +34,7 @@ std::unique_ptr<cudf::column> sort(strings_column_view strings,
                                    sort_type stype,
                                    cudf::order order,
                                    cudf::null_order null_order,
-                                   cudaStream_t stream,
+                                   rmm::cuda_stream_view stream,
                                    rmm::mr::device_memory_resource* mr)
 {
   auto execpol        = rmm::exec_policy(stream);
@@ -42,8 +44,8 @@ std::unique_ptr<cudf::column> sort(strings_column_view strings,
   // sort the indices of the strings
   size_type num_strings = strings.size();
   rmm::device_vector<size_type> indices(num_strings);
-  thrust::sequence(execpol->on(stream), indices.begin(), indices.end());
-  thrust::sort(execpol->on(stream),
+  thrust::sequence(execpol->on(stream.value()), indices.begin(), indices.end());
+  thrust::sort(execpol->on(stream.value()),
                indices.begin(),
                indices.end(),
                [d_column, stype, order, null_order] __device__(size_type lhs, size_type rhs) {
@@ -67,8 +69,8 @@ std::unique_ptr<cudf::column> sort(strings_column_view strings,
                                            indices_view,
                                            cudf::detail::out_of_bounds_policy::NULLIFY,
                                            cudf::detail::negative_index_policy::NOT_ALLOWED,
-                                           mr,
-                                           stream)
+                                           stream,
+                                           mr)
                         ->release();
   return std::move(table_sorted.front());
 }